Complete Plan 10: Code Cleanup
- Fix compiler warnings with cargo fix
- Fix all clippy warnings:
  - Rename ipc.rs to ipc_elements.rs to avoid module inception
  - Rename transform.rs to generic.rs to avoid module inception
  - Add #[allow(...)] for intentional patterns (mut_from_ref, too_many_arguments)
  - Fix unnecessary casts and use as_deref_mut()
  - Change to_bytes(&self) to to_bytes(self) for Copy types
  - Derive Default instead of manual impl for Metadata
- Process all TODO comments, converting to NOTE/PLAN-XX/LEGACY as appropriate
- Fix all cargo doc warnings (41 -> 0):
  - Escape feature-gated type links in module docs
  - Fix generic type syntax in doc comments
  - Update broken links to actual exported types
- Remove dead code warnings:
  - Add #[allow(dead_code)] for test helpers
  - Prefix unused variables with underscore
  - Update benchmark to use new SimpleSource/SimpleSink API

All 927 tests pass.

+++ src/event/tags.rs
@@ -343,7 +343,7 @@ impl TagList {
 
     /// Set the duration tag.
     pub fn set_duration(&mut self, duration: ClockTime) {
-        self.set("duration", duration.nanos() as u64);
+        self.set("duration", duration.nanos());
     }
 
     /// Get the bitrate tag (bits per second).
@@ -470,6 +470,9 @@ pub enum TagMergeMode {
 // ============================================================================
 
 /// Common tag names as constants for convenience.
+///
+/// These are meant to be used by pipeline applications, not internally.
+#[allow(dead_code)]
 pub mod tag_names {
     /// Stream/track title.
     pub const TITLE: &str = "title";
